JAVA_FSD

+++ RLL-Backend/EmployeePayrollManagement/src/main/java/com/attendance/demo/repo/AttendanceRepo.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.attendance.demo.entity.Attendance;
 
-public interface AttendanceRepo extends JpaRepository<Attendance,Integer>{
+public interface AttendanceRepo extends JpaRepository<Attendance,Long>{
 	
 
 }
